requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cmudict==1.0.32
cn2an==0.5.23
discord.py==2.5.2
Distance==0.1.3
docopt==0.6.2
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
g2p-en==2.1.0
huggingface-hub==0.31.2
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflect==7.5.0
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.0
llvmlite==0.44.0
loguru==0.7.3
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
nltk==3.9.1
num2words==0.5.14
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
proces==0.1.7
propcache==0.3.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyNaCl==1.5.0
pyopenjtalk-dict==0.3.4.dev2
pypinyin==0.54.0
python-dotenv==1.1.0
pyworld-prebuilt==0.3.5.post1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
starlette==0.46.2
style-bert-vits2==2.5.0
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
typeguard==4.4.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
zipp==3.21.0
